Fixed Window Resize Bug
Fixed the bug where window would always resize to the set value. A
consequence of this is that the application begins in max screen size
(not necessarily a bad thing) every time, however, this can be adjusted.
It should work for all resolutions though and will not resize if the
screen is minimized and you switch between screens.